Minor tidy ups
